IT WORKS
i had to reset the db bc for some reason it wasnt making a table for chat_conversations???? anyways a recruiter (and only a recruiter) can start a convo with a seeker they dont already have a conversation with. after that seekers and recruiters can send messages. if youre not one of the 2 users then you just get booted back to list of convos